+++ src/controllers/sites.js
@@ -22,7 +22,7 @@ import {
   hasText,
   isBoolean,
   isObject,
-  getStoredMetrics, getRUMDomainKey, isValidUUID, deepEqual,
+  getStoredMetrics, isValidUUID, deepEqual,
 } from '@adobe/spacecat-shared-utils';
 import { Site as SiteModel } from '@adobe/spacecat-shared-data-access';
 
@@ -401,15 +401,12 @@ function SitesController(dataAccess, log, env) {
     const domain = wwwUrlResolver(site);
 
     try {
-      const domainKey = await getRUMDomainKey(site.getBaseURL(), context);
       const current = await rumAPIClient.query(TOTAL_METRICS, {
         domain,
-        domainkey: domainKey,
         interval: MONTH_DAYS,
       });
       const total = await rumAPIClient.query(TOTAL_METRICS, {
         domain,
-        domainkey: domainKey,
         interval: 2 * MONTH_DAYS,
       });
       const organicTraffic = await getStoredMetrics(
@@ -440,11 +437,7 @@ function SitesController(dataAccess, log, env) {
         projectedTrafficValue,
       });
     } catch (error) {
-      if (error.message?.includes('Error retrieving the domain key')) {
-        log.info(`No RUM key configured for site ${siteId}: ${error.message}`);
-      } else {
-        log.error(`Error getting RUM metrics for site ${siteId}: ${error.message}`);
-      }
+      log.error(`Error getting RUM metrics for site ${siteId}: ${error.message}`);
     }
 
     return ok({
@@ -452,6 +445,41 @@ function SitesController(dataAccess, log, env) {
       ctrChange: 0,
       projectedTrafficValue: 0,
     });
+  };
+
+  const getPageMetricsBySource = async (context) => {
+    const siteId = context.params?.siteId;
+    const metric = context.params?.metric;
+    const source = context.params?.source;
+    const encodedPageURL = context.params?.base64PageUrl;
+
+    if (!isValidUUID(siteId)) {
+      return badRequest('Site ID required');
+    }
+
+    if (!hasText(metric)) {
+      return badRequest('metric required');
+    }
+
+    if (!hasText(source)) {
+      return badRequest('source required');
+    }
+
+    if (!hasText(encodedPageURL)) {
+      return badRequest('base64PageUrl required');
+    }
+
+    const decodedPageURL = Buffer.from(encodedPageURL, 'base64').toString('utf-8').trim();
+
+    const site = await Site.findById(siteId);
+    if (!site) {
+      return notFound('Site not found');
+    }
+
+    let metrics = await getStoredMetrics({ siteId, metric, source }, context);
+    metrics = metrics.filter((metricEntry) => metricEntry.url === decodedPageURL);
+
+    return ok(metrics);
   };
 
   return {
@@ -474,6 +502,7 @@ function SitesController(dataAccess, log, env) {
 
     // site metrics
     getSiteMetricsBySource,
+    getPageMetricsBySource,
     getLatestSiteMetrics,
   };
 }