+++ CHANGELOG.md
@@ -1,3 +1,66 @@
+# [1.90.0](https://github.com/adobe/spacecat-api-service/compare/v1.89.18...v1.90.0) (2025-02-13)
+
+
+### Features
+
+* **page-metrics:** API to get page level metrics by page URL ([#742](https://github.com/adobe/spacecat-api-service/issues/742)) ([3b03a40](https://github.com/adobe/spacecat-api-service/commit/3b03a40876d44550fcbac73f1b3b425e466d8a31))
+
+## [1.89.18](https://github.com/adobe/spacecat-api-service/compare/v1.89.17...v1.89.18) (2025-02-12)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-data-access to v2.2.0 ([#746](https://github.com/adobe/spacecat-api-service/issues/746)) ([f247eeb](https://github.com/adobe/spacecat-api-service/commit/f247eeb105785d9165a2ec719137578c08e9a96e))
+
+## [1.89.17](https://github.com/adobe/spacecat-api-service/compare/v1.89.16...v1.89.17) (2025-02-12)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-utils to v1.31.0 ([#745](https://github.com/adobe/spacecat-api-service/issues/745)) ([3994d5d](https://github.com/adobe/spacecat-api-service/commit/3994d5de2960f5de91cc3680b4f9a97866a12bea))
+
+## [1.89.16](https://github.com/adobe/spacecat-api-service/compare/v1.89.15...v1.89.16) (2025-02-12)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#740](https://github.com/adobe/spacecat-api-service/issues/740)) ([2b03d50](https://github.com/adobe/spacecat-api-service/commit/2b03d505ac3364a62394deb3f6905d0a7e7977d4))
+
+## [1.89.15](https://github.com/adobe/spacecat-api-service/compare/v1.89.14...v1.89.15) (2025-02-11)
+
+
+### Bug Fixes
+
+* **rum-api-client:** change to new rum api client ([#741](https://github.com/adobe/spacecat-api-service/issues/741)) ([36cb41a](https://github.com/adobe/spacecat-api-service/commit/36cb41a1afd35c1ae36f12f41bafa720d9d68f09))
+
+## [1.89.14](https://github.com/adobe/spacecat-api-service/compare/v1.89.13...v1.89.14) (2025-02-09)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#737](https://github.com/adobe/spacecat-api-service/issues/737)) ([7c428a7](https://github.com/adobe/spacecat-api-service/commit/7c428a769ebbfbd551aae3ad81252a2b31dea0f7))
+
+## [1.89.13](https://github.com/adobe/spacecat-api-service/compare/v1.89.12...v1.89.13) (2025-02-08)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#736](https://github.com/adobe/spacecat-api-service/issues/736)) ([1500e4a](https://github.com/adobe/spacecat-api-service/commit/1500e4ab1f52e1317081e8ac5a5263948df22033))
+
+## [1.89.12](https://github.com/adobe/spacecat-api-service/compare/v1.89.11...v1.89.12) (2025-02-08)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/spacecat-shared-rum-api-client to v2.20.1 ([#735](https://github.com/adobe/spacecat-api-service/issues/735)) ([edf4bd7](https://github.com/adobe/spacecat-api-service/commit/edf4bd7cc167218381768149cba9c753ba042708))
+
+## [1.89.11](https://github.com/adobe/spacecat-api-service/compare/v1.89.10...v1.89.11) (2025-02-06)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#734](https://github.com/adobe/spacecat-api-service/issues/734)) ([6172aa8](https://github.com/adobe/spacecat-api-service/commit/6172aa8119895b11425dd0c6f7d2103c02ec8ed8))
+
 ## [1.89.10](https://github.com/adobe/spacecat-api-service/compare/v1.89.9...v1.89.10) (2025-02-03)